Refactored styles and added getDayOfWeek and getMonth function
Refactored styles into separate files and added more time helpers for global use.
