+++ code/TimeManagementApp/pages/css/createGoalsStyles.js
@@ -1,0 +1,41 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'stretch',
+    justifyContent: 'flex-start',
+    backgroundColor: '#20639B'
+  },
+  sectionTitleContainer: {
+    alignItems: 'center',
+  },
+  sectionTitle: {
+    fontSize: 32,
+    textDecorationLine: 'underline',
+    color: 'white'
+  },
+  sectionContainer: {
+     flex: 4,
+     alignItems: 'stretch'
+  },
+  formContainer: {
+    flex: 1,
+    marginLeft: 10,
+    marginRight: 20,
+  },
+  formInput: {
+    marginTop: 10,
+    backgroundColor: '#D3D3D3'
+  },
+  formDateInput: {
+    marginTop: 10,
+    backgroundColor: 'white'
+  },
+  timeBtn: {
+    marginTop: 5
+  },
+  createContainer: {
+    marginTop: 40
+  }
+});
