+++ code/TimeManagementApp/pages/home.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import {
   Button,
   SafeAreaView,
-  StyleSheet,
   ScrollView,
   View,
   Text,
@@ -24,87 +23,11 @@ import {
 import {NavigationContainer} from '@react-navigation/native';
 import AsyncStorage from '@react-native-community/async-storage';
 
-import { getGoals, clearGoals } from '../logic/goals.js';
-import { convertTo12HourFormat } from '../helpers/timeHelper.js';
+// Import stylesheet
+import styles from './css/homeStyles.js';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'stretch',
-    justifyContent: 'flex-start',
-    backgroundColor: '#20639B'
-  },
-  quoteContainer: {
-    flex: 1,
-    backgroundColor: 'black',
-    alignItems: 'stretch',
-    paddingLeft: 10,
-    paddingRight: 10
-  },
-  quote: {
-    color: 'white',
-    fontStyle: 'italic',
-    fontSize: 16
-  },
-  sectionContainer: {
-     flex: 4,
-     alignItems: 'stretch'
-  },
-  sectionHeader: {
-    flex: 0.5,
-    flexDirection: 'row',
-    alignItems: 'flex-start'
-  },
-  sectionTitle: {
-    marginLeft: 10,
-    fontSize: 32,
-    textDecorationLine: 'underline',
-    color: 'white'
-  },
-  tableContainer: {
-    flex: 3,
-    marginLeft: 10,
-    marginRight: 20,
-  },
-  sectionTable: {
-    flex: 1,
-    alignItems: "stretch",
-    backgroundColor: 'white'
-  },
-  tableHeader: {
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start',
-    borderTopWidth: 0,
-    borderLeftWidth: 0,
-    borderRightWidth: 0,
-    borderBottomWidth: 1,
-  },
-  rowItem: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center'
-  },
-  tableRowEmpty: {
-    height: 50,
-    alignItems: 'center',
-    justifyContent: 'flex-start'
-  },
-  tableRowEven: {
-    height: 50,
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start'
-  },
-  tableRowOdd: {
-    height: 50,
-    backgroundColor: '#D3D3D3',
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start'
-  },
-});
+import { getGoals, clearGoals } from '../logic/goals.js';
+import { convertTo12HourFormat, formatDate } from '../helpers/timeHelper.js';
 
 export default function HomeScreen({ navigation }) {
   const [quoteExpanded, expandQuote] = useState(false);
@@ -199,7 +122,7 @@ function goalsTable() {
                     </View>
                     <View style={styles.rowItem}>
                       <Text>{convertTo12HourFormat(item.date)}</Text>
-                      <Text>{item.date.getDate() + "/" + item.date.getMonth() + "/" + item.date.getFullYear()}</Text>
+                      <Text>{formatDate(item.date)}</Text>
                     </View>
                   </View>
                 );
